Refactor claims api swagger docs Upgrade claims api swagger docs to oas3
