+++ modules/claims_api/app/controllers/claims_api/docs/api_controller.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require_dependency 'claims_api/form_526_model_swagger'
-require_dependency 'claims_api/form_0966_model_swagger'
-require_dependency 'claims_api/form_2122_model_swagger'
+require_dependency 'claims_api/forms/form_526_response_swagger'
+require_dependency 'claims_api/forms/form_0966_response_swagger'
+require_dependency 'claims_api/forms/form_2122_response_swagger'
 
 module ClaimsApi
   module Docs
