+++ modules/claims_api/app/swagger/claims_api/v0/security_scheme_swagger.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module ClaimsApi
+  module V0
+    class SecuritySchemeSwagger
+      include Swagger::Blocks
+      swagger_component do
+        security_scheme :apikey do
+          key :type, :apiKey
+          key :name, :apikey
+          key :in, :header
+        end
+      end
+    end
+  end
+end
